backend/pyproject.toml
[project]
name = "triad-backend"
version = "0.1.0"
description = "TRIAD — Multi-Agent Consensus System Backend"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=13.0",
    "python-multipart>=0.0.9",

    # LLM Providers
    "openai>=1.50.0",
    "anthropic>=0.34.0",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",

    # Neo4j
    "neo4j>=5.20.0",
    "neo4j-graphrag>=1.0.0",

    # Data & Config
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",

    # Utilities
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
